TODO-1 feat: Add and implement error handling

+++ src/tasks/services/tasks.service.ts
@@ -9,9 +9,10 @@ import { UsersEntity } from 'src/users/entities/users.entity';
 import { TASKS_STATUS } from 'src/constants/TASKS_STATUS';
 import { updateTaskStatusDTO } from '../dto/update-task-status.dto';
 import { IUseToken } from 'src/auth/interfaces/auth.interface';
-import { manageTokenFromHeaders } from 'src/utils/manage.token';
+import { manageTokenFromHeaders } from 'src/utils/token.manager';
 import { ITasksService } from 'src/interfaces/tasks/tasks-service.interface';
 import { IUsersService } from 'src/interfaces/users/users-service.interface';
+import { ErrorManager } from 'src/utils/error.manager';
 
 @Injectable()
 export class TasksService implements ITasksService {
@@ -26,44 +27,52 @@ export class TasksService implements ITasksService {
     // Methods by Id
 
     public async getAllTasksByUserId(userId: string): Promise<TasksEntity[]> {
-        const user: UsersEntity = await this.usersService.getUserById(userId);
 
-        const tasks: TasksEntity[] = await this.tasksRepository
-            .findBy({
-                user
-            })
+        try {
+            const user: UsersEntity = await this.usersService.getUserById(userId);
+    
+            const tasks: TasksEntity[] = await this.tasksRepository
+                .findBy({
+                    user
+                });
+    
+            return tasks;
 
-        return tasks;
+        } catch(error) {
+            throw ErrorManager.createSignatureError(error.message);
+        }
     }
 
     public async getTaskById(taskId: string): Promise<TasksEntity> {
         try {
-            const task = this.tasksRepository.findOneBy({
+            const task = await this.tasksRepository.findOneBy({
                 taskId
-            })
+            });
 
-            if (task) {
-                return task
+            if (!task) {
+                throw new ErrorManager({
+                    type: 'NOT_FOUND',
+                    message: 'The task was not found'
+                });
             }
-
-            throw new NotFoundException();
+            
+            return task;
 
         } catch(error) {
-            throw new NotFoundException();
+            throw ErrorManager.createSignatureError(error.message);
         }
         
     }
 
     public async updateTaskStatusById(taskId: string, body: updateTaskStatusDTO): Promise<UpdateResult> {
         try {
-            const task = await this.getTaskById(taskId)
-
-            if (!task) {
-                throw new BadRequestException();
-            }
+            const task: TasksEntity = await this.getTaskById(taskId);
             
             if (body.status !== task.status) {
-                throw new BadRequestException();
+                throw new ErrorManager({
+                    type: 'BAD_REQUEST',
+                    message: 'The current status of the task does not match with the db'
+                });
             }
 
             if (body.status === TASKS_STATUS.PENDING) {
@@ -73,52 +82,68 @@ export class TasksService implements ITasksService {
             }
             const updatedTask: UpdateResult = await this.tasksRepository.update(
                 taskId, task
-            )
+            );
 
-            if (updatedTask.affected === 1) {
-                return updatedTask;
+            if (updatedTask.affected === 0) {
+                throw new ErrorManager({
+                    type: 'NOT_MODIFIED',
+                    message: 'The task status could not be updated'
+                });
             }
-            throw new BadRequestException();
+
+            return updatedTask;
+
         } catch(error) {
-            throw new BadRequestException();
+            throw ErrorManager.createSignatureError(error.message);
         }
 
     }
 
     public async deleteTaskById(taskId: string): Promise<DeleteResult> {
         try {
-            const task: TasksEntity = await this.getTaskById(taskId);
+            await this.getTaskById(taskId);
 
             const deleteResult: DeleteResult = await this.tasksRepository.delete(
-                task.taskId
-            )
+                taskId
+            );
 
-            if (deleteResult.affected !== 0) {
-                return deleteResult;
+            if (deleteResult.affected === 0) {
+                throw new ErrorManager({
+                    type: 'BAD_REQUEST',
+                    message: 'The task could not be deleted'
+                });
             }
-
-            throw new BadRequestException();
+            
+            return deleteResult;
 
         } catch(error) {
-            throw new BadRequestException();
+            throw ErrorManager.createSignatureError(error.message);
         }
     }
 
-    public async deleteAllTasksByUserId(userId: string): Promise<DeleteResult> {
+    public async deleteAllTasksByUserId(userId: string): Promise<DeleteResult | void> {
         try {
+
+            if ((await this.getAllTasksByUserId(userId)).length === 0) {
+                return;
+            }
+
             const deleteResult: DeleteResult = await this.tasksRepository.createQueryBuilder('tasks')
                 .delete()
                 .where('user.userId = :userId', {userId})
                 .execute();
 
-            if (deleteResult.affected !== 0) {
-                return deleteResult;
+            if (deleteResult.affected === 0) {
+                throw new ErrorManager({
+                    type: 'BAD_REQUEST',
+                    message: 'The user tasks could not be deleted'
+                });
             }
-
-            throw new BadRequestException();
+            
+            return deleteResult;
 
         } catch(error) {
-            throw new BadRequestException();
+            throw ErrorManager.createSignatureError(error.message);
         }
     }
 
@@ -134,18 +159,14 @@ export class TasksService implements ITasksService {
             const managedToken: IUseToken = manageTokenFromHeaders(request);
             
             // sub refers to Subject => userId
-            const tokenUserId = managedToken.sub
+            const tokenUserId = managedToken.sub;
             
-            const tasks: TasksEntity[] = await this.getAllTasksByUserId(tokenUserId)
-
-            if (tasks.length === 0) {
-                throw new NotFoundException()
-            }
+            const tasks: TasksEntity[] = await this.getAllTasksByUserId(tokenUserId);
 
             return tasks;
 
         } catch(error) {
-            console.log(error)
+            throw ErrorManager.createSignatureError(error.message);
         }
     }
 
@@ -166,7 +187,7 @@ export class TasksService implements ITasksService {
             })
 
         } catch(error) {
-            throw new BadRequestException();
+            throw ErrorManager.createSignatureError(error.message);
         }
     }
 }
